--- client/app.scss
@@ -1,36 +1,0 @@
-$fa-font-path:  "./fonts" !default;
-@import '../node_modules/font-awesome/scss/font-awesome';
-@import '../node_modules/substance/styles/base/index';
-
-// Substance modules
-@import '../node_modules/substance/styles/components/_all';
-@import '../node_modules/substance/packages/_all';
-
-// Notes
-@import '../packages/notes/_notes'; // app styles
-@import '../packages/notes/_shared'; // buttons etc.
-@import '../packages/header/_header';
-@import '../packages/dashboard/_dashboard';
-@import '../packages/dashboard/_note-item';
-@import '../packages/notes/_request-login';
-@import '../packages/notes/_enter-name';
-@import '../packages/header/_login-status';
-@import '../packages/notes/_welcome';
-@import '../packages/collaborators/_collaborators';
-@import '../packages/notification/_notification';
-
-// NotesWriter styles
-@import '../packages/note-writer/_note-writer';
-@import '../packages/note-cover/_note-cover';
-@import '../packages/note-cover/_note-summary';
-@import '../packages/todo/_todo';
-@import '../packages/mark/_mark';
-@import '../packages/comment/_comment';
-
-
-
-/* Shared demo styles */
-
-body.sm-fixed-layout {
-  overflow: hidden;
-}
