+++ client/index.html
@@ -1,13 +1,11 @@
-<!DOCTYPE html>
 <html>
-  <head>
-    <title>Substance Notepad</title>
-    <meta charset="UTF-8">
-    <!--CONFIG-->
-    <link href='app.css' rel='stylesheet' type='text/css'/>
-    <script src='app.js'></script>
-  </head>
-  <body>
-    <!-- app goes here -->    
-  </body>
+<head>
+  <title>Substance Notes</title>
+  <meta charset="UTF-8">
+  <script type="text/javascript" src="./app.js"></script>
+  <link href='./app.css' rel='stylesheet' type='text/css'/>
+</head>
+<body>
+
+</body>
 </html>
